[#34561] Fix: `@8110' is not allowed as an instance variable name
Some entries in the syntax dictionary does not have a title. In those
cases the AI is used as the title. The title is used to set the class
name. The class name is later used as the name of an instance
variable. Instance variable names cannot start with a number. All AIs
are numeric. The solution is to prepend `_` if the class name starts
with a number.